FieldConfigurationHelper updates: interface and index-only fields
 . Extracted FieldConfigHelper interface
 . Implementation of FieldConfigurationHelper is XMLFieldConfigHelper
 . FieldConfigHelper now used to configure index-only fields.
 . index-only fields are store=false and indexed=true
